requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==39.0.2
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.53.1
django-bootstrap-v5==1.0.11
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-database-url==1.0.3
django-environ==0.10.0
django-filter==22.1
django-forms-bootstrap==3.1.0
django-js-asset==2.0.0
django_debug_toolbar==3.8.1
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
whitenoise==6.4.0
